requirements.txt
asgiref==3.7.2
Django==4.2.2
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary
djangorestframework